pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-migrate>=4.1.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "supabase>=2.18.1",
    "python-dotenv>=1.1.1",
]
